Refactor attribute rendering boundaries
